Add test for suffix arrays
The fast algorithm for creating a suffix array (SA) has a bug, and the latest
test consistently fails. The problem lies in the incorrect placement of "2" in
the first position and "0" in the second position by the fast algorithm (eg). I
have conducted some initial investigation, but I have not identified the root cause
of the issue. I have identified a section of code that might be responsible
for the problem, but further study is required to pinpoint and resolve the bug.